requirements.txt
alabaster==0.7.12
autopep8==1.6.0
Babel==2.10.1
black==21.12b0
certifi==2022.5.18.1
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==36.0.2
distlib==0.3.4
docutils==0.17.1
filelock==3.7.1
Flask==2.1.1
Flask-JWT==0.3.2
Flask-MySQLdb==1.0.1
identify==2.5.1
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
mysqlclient==2.1.0
nodeenv==1.6.0
packaging==21.3
Pallets-Sphinx-Themes==2.0.2
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.18.1
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.28.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
toml==0.10.2
tomli==1.2.3
tox==3.25.0
typing_extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
Werkzeug==2.1.2
zipp==3.8.0
